# Competition between patterns to solve the comment/string/regexp problem