--- a/components/prism-javascript.js
+++ b/components/prism-javascript.js
@@ -6,7 +6,8 @@ Prism.languages.javascript = Prism.languages.extend('clike', {
 Prism.languages.insertBefore('javascript', 'keyword', {
 	'regex': {
 		pattern: /(^|[^/])\/(?!\/)(\[.+?]|\\.|[^/\r\n])+\/[gim]{0,3}(?=\s*($|[\r\n,.;})]))/g,
-		lookbehind: true
+		lookbehind: true,
+		competition: 'enclosing'
 	}
 });