@Golmote 2015-01-03
Hi there!

I thought a lot about this stuff during the holidays. Your suggestion using `start` and `end` patterns would raise some tricky problems...
To match the whole symbol, we'd need to merge the `start`, `contents` and `end` regexps together, leading to the same problems encountered in PR #444! What happens if the user sets the `i` modifier only on the `start` token, for example?

What do you think? Can we sort it out or should we go back to the boolean solution, and just find a better name than `competition`?