@zeitgeist87 2015-01-03
You don't have to combine them. You search for `start` and get an index, then you search for `end` and get an index, then you check if `contents` matches inbetween. If it's possible, you combine them of course.