@zeitgeist87 2015-01-03
You could even optimize a bit, by matching only `start` for tokens that are in competition. Then sort the resulting list of start indices in decending order. Then go through the list and match `end` and `contents`. The first match is the new token.