@Golmote 2015-01-03
This looked like an awesome idea until I tested strings like `"foo \" bar"`. When searching for `end`, we find the escaped quote.
Even if we add some lookbehind check to ensure the quote is not escaped, it will fail on `"foo \\"` by thinking the quote is escaped when it's not...

So we might need to consume `contents` before searching for `end`, but this would require `contents` to be more precise. Like `/(\\.|[^"])*/` instead of just `/(\\?.)*/` for strings, for example. This is probably not easy enough for "a random dev, new to Prism, trying to hack together a language definition".

What do you think?