@zeitgeist87 2015-01-03
@Golmote yes you are absolutely right. I hadn't thought of escaped end patterns. You could also not use back references to refer to the string that was matched with `start`.

A `boolean` flag `competition` would probably be best, although I like your idea of allowing different scopes of competition by nameing them with a string.