@LeaVerou 2015-01-04
@Golmote: If the dev gives us the escape character too (which would default to `\`, but other languages may have different escape characters), can’t we build the necessary regex ourselves in the background?

I really think the flag idea is confusing for devs trying to hack together language definitions. Prism’s priorities are: 1. API & code simplicity 2. Performance 3. Small code size, in that order.

Also, have we done any performance comparison between this compared to previous versions of Prism that failed on such nested structures? A small performance hit is acceptable, but if Prism has to become e.g. 4 times slower to become a bit more correct, it wouldn't be worth it. It's a tradeoff.