@Golmote 2015-01-04
I just ran a quick benchmark on the three concerned versions of Prism: the original (current state on Prism website), the one from this PR (in the current state) and the one from PR #444 (at commit zeitgeist87/prism@0187761a56d8abdc83288921e4c4782ecbab955d).

I did not benchmark your proposed API as I could not yet make it work properly.

The benchmark highlights the following code:
 ```javascript
/*fo"/o*/ "*/bar", /ba"z*/;
"Foo // bar"
"/foo/"; /* "bar" /baz/ */
/"foo"/, "/*bar", "baz*/";
/foo"/, "bar", "baz";
// foo "bar" baz
"foo /* bar */ baz";
```

I repeated this code fragment multiple times to test for longer codes.

The code was highlighted a first time before the benchmark runs so that I could check the result and ignore the grammar optimization time in PR #444.

Here are the results in Chrome 39.0.2171.95:

Prism version | Code length | Number of runs | Average time
---|---|---|---
Original | 155 chars | 1000 | 0.29 ms
PR #424 | 155 chars | 1000 | 0.37 ms (27% slower)
PR #444 | 155 chars | 1000 | 0.35 ms (20% slower)
Original | 1550 chars | 100 | 1.86 ms
PR #424 | 1550 chars | 100 | 2.64 ms (41% slower)
PR #444 | 1550 chars | 100 | 2.38 ms (28% slower)
Original | 15500 chars | 50 | 9.99 ms
PR #424 | 15500 chars | 50 | 14.28 ms (43% slower)
PR #444 | 15500 chars | 50 | 13.67 ms (37% slower)
Original | 155000 chars | 10 | 171.87 ms
PR #424 | 155000 chars | 10 | 208.23 ms (21% slower)
PR #444 | 155000 chars | 10 | 200.53 ms (17% slower)

Here are the results in Firefox 34.0.5:

Prism version | Code length | Number of runs | Average time
---|---|---|---
Original | 155 chars | 1000 | 1.11 ms
PR #424 | 155 chars | 1000 | 1.46 ms (31% slower)
PR #444 | 155 chars | 1000 | 1.39 ms (25% slower)
Original | 1550 chars | 100 | 5.53 ms
PR #424 | 1550 chars | 100 | 7.08 ms (28% slower)
PR #444 | 1550 chars | 100 | 7.12 ms (28% slower)
Original | 15500 chars | 50 | 51.35 ms
PR #424 | 15500 chars | 50 | 74.60 ms (45% slower)
PR #444 | 15500 chars | 50 | 69.39 ms (35% slower)
Original | 155000 chars | 10 | 1022.95 ms
PR #424 | 155000 chars | 10 | 1705.92 ms (66% slower)
PR #444 | 155000 chars | 10 | 2153.13 ms (110% slower)

Finally, here is the benchmark code I used:

```html
<pre><code id="benchmark-core" class="language-javascript">/*fo"/o*/ "*/bar", /ba"z*/;
"Foo // bar"
"/foo/"; /* "bar" /baz/ */
/"foo"/, "/*bar", "baz*/";
/foo"/, "bar", "baz";
// foo "bar" baz
"foo /* bar */ baz";
</code></pre>

<script src="components/prism-core.js"></script>
<script src="components/prism-clike.js"></script>
<script src="components/prism-javascript.js"></script>
<script>
	(function() {
		var elt = document.getElementById('benchmark-core'),
			code = elt.textContent;

		Prism.highlightElement(elt);

		function avg(arr) {
			return arr.reduce(function(a,b) { return a+b; })/arr.length;
		}
		function str_repeat(s, n) {
			var res = '';
			while(n--) {
				res += s;
			}
			return res;
		}

		function benchmark(code, cpt, cb) {
			var logs = [],
				f = function() {
					if(cpt--) {
						var start = performance.now();
						Prism.highlight(code, Prism.languages.javascript, 'javascript');
						logs.push(performance.now() - start);
						requestAnimationFrame(f);
					} else {
						console.log('Average time: ' + avg(logs) + ' ms');
						cb && cb();
					}
				};
			console.log('Highlighting '+code.length+'-char code '+cpt+' times.');
			f();
		}

		window.onload = function() {
			benchmark(code, 1000, function () {
				benchmark(str_repeat(code, 10), 100, function () {
					benchmark(str_repeat(code, 100), 50, function () {
						benchmark(str_repeat(code, 1000), 10, function () {
							console.log('Done');
						});
					});
				});
			});
		};
	}());
</script>
```